Simplify code using new hook useLoadAssignments to return most state going to be used to display assignment selection and review
