+++ src/containers/Review/assignment/AssignmentSubmit.tsx
@@ -10,7 +10,6 @@ interface AssignmentSubmitProps {
   assignmentsFiltered: AssignmentDetails[]
   enableSubmit: boolean
   setAssignmentError: React.Dispatch<SetStateAction<string | null>>
-  refetch: () => void
 }
 
 const AssignmentSubmit: React.FC<AssignmentSubmitProps> = ({
@@ -18,8 +17,7 @@ const AssignmentSubmit: React.FC<AssignmentSubmitProps> = ({
   assignedSectionsByLevel,
   assignmentsFiltered,
   enableSubmit,
-  setAssignmentError,
-  refetch,
+  setAssignmentError
 }) => {
   const { strings } = useLanguageProvider()
   const { submitAssignments } = useUpdateAssignment({
@@ -30,7 +28,7 @@ const AssignmentSubmit: React.FC<AssignmentSubmitProps> = ({
     try {
       for (const [level, assignedSections] of Object.entries(assignedSectionsByLevel)) {
         await submitAssignments(Number(level), assignedSections, assignmentsFiltered)
-        refetch()
+        fullStructure.reload()
       }
     } catch (err) {
       console.log(err)
