+++ src/containers/Review/assignment/AssignmentTab.tsx
@@ -1,81 +1,54 @@
-import React, { useEffect, useState } from 'react'
-import { Container, Label, Message } from 'semantic-ui-react'
-import Loading from '../../../components/Loading'
-import { useUserState } from '../../../contexts/UserState'
+import React, { useState } from 'react'
+import { Container, Header, Label, Message, Segment } from 'semantic-ui-react'
 import {
   AssignedSectionsByLevel,
   AssignmentDetails,
   Filters,
   FullStructure,
-  LevelAssignments,
+  SectionDetails,
 } from '../../../utils/types'
 import { useLanguageProvider } from '../../../contexts/Localisation'
 import { Stage } from '../../../components/Review'
-import { getPreviousStageAssignment } from '../../../utils/helpers/assignment/getPreviousStageAssignment'
-import useGetReviewInfo from '../../../utils/hooks/useGetReviewInfo'
 import { NoMatch } from '../../../components'
-import ReviewLevel, { ALL_LEVELS } from './ReviewLevel'
-import { ReviewStateProvider } from '../../../contexts/ReviewStructuresState'
-import AssignmentRows from './AssignmentRows'
+import Loading from '../../../components/Loading'
+import ReviewLevel from './ReviewLevel'
+import {
+  ReviewStateProvider,
+  useReviewStructureState,
+} from '../../../contexts/ReviewStructuresState'
 import AssignmentSubmit from './AssignmentSubmit'
 import AssignAll from './AssignAll'
 import { ApplicationOutcome } from '../../../utils/generated/graphql'
+import useLoadAssignments from '../../../utils/hooks/useLoadAssignments'
+import AssignmentSectionRow from './AssignmentSectionRow'
+import ReviewSectionRow from './ReviewSectionRow'
 
 const AssignmentTab: React.FC<{
   fullApplicationStructure: FullStructure
 }> = ({ fullApplicationStructure: fullStructure }) => {
+  const sectionCodes = Object.keys(fullStructure.sections)
   const { strings } = useLanguageProvider()
-  const {
-    userState: { currentUser },
-  } = useUserState()
   const [enableSubmit, setEnableSubmit] = useState<boolean>(false)
   const [assignedSectionsByLevel, setAssignedSectionsByLevel] = useState<AssignedSectionsByLevel>(
     {}
   )
-
   const [filters, setFilters] = useState<Filters | null>(null)
   const [assignmentError, setAssignmentError] = useState<string | null>(null)
-  const [assignmentsFiltered, setAssignmentsFiltered] = useState<AssignmentDetails[]>([])
 
-  const { error, loading, assignments, refetch } = useGetReviewInfo({
-    applicationId: fullStructure.info.id,
-    serial: fullStructure.info.serial,
-    userId: currentUser?.userId as number,
+  const {
+    error,
+    loading,
+    assignmentsFiltered,
+    assignmentInPreviousStage,
+    assignmentGroupedLevel,
+    isFullyAssigned,
+  } = useLoadAssignments({
+    info: fullStructure.info,
+    sectionCodes,
+    filters,
   })
 
-  useEffect(() => {
-    if (assignments) {
-      const filteredByLevel = getFilteredLevel(assignments)
-      setAssignmentsFiltered(filteredByLevel)
-    }
-  }, [filters, assignments])
-
-  const getFilteredByStage = (assignments: AssignmentDetails[]) => {
-    if (!filters) return []
-    return assignments
-      .filter(({ isLocked }) => !isLocked) // Should not give option to assign if Assignment is locked
-      .filter((assignment) => assignment.current.stage.number === filters.currentStage)
-  }
-
-  const getFilteredLevel = (assignments: AssignmentDetails[]) => {
-    if (!filters) return []
-    return getFilteredByStage(assignments).filter(
-      (assignment) =>
-        filters.selectedLevel === ALL_LEVELS || assignment.level === filters.selectedLevel
-    )
-  }
-
   if (error) return <Message error header={strings.ERROR_REVIEW_PAGE} list={[error]} />
-
-  if (loading || !fullStructure) return <Loading />
-  if (!assignments || assignments.length === 0) return <NoMatch />
-
-  // Get Previous stage (last level reviewer) assignment
-  const assignmentInPreviousStage = getPreviousStageAssignment(
-    fullStructure.info.serial,
-    assignments,
-    filters?.currentStage
-  )
 
   const {
     info: {
@@ -84,8 +57,6 @@ const AssignmentTab: React.FC<{
       },
     },
   } = fullStructure
-
-  const sectionCodes = Object.keys(fullStructure.sections)
 
   const assignAllSections = (reviewerId: number) => {
     const alreadyAssignedSections = new Set(
@@ -103,21 +74,59 @@ const AssignmentTab: React.FC<{
       if (!alreadyAssignedSections.has(section))
         newAssignments[section] = { newAssignee: reviewerId }
     })
+
+    const currentReviewLevel = Math.max(Number(Object.keys(assignmentGroupedLevel)))
     setAssignedSectionsByLevel({ ...assignedSectionsByLevel, [currentReviewLevel]: newAssignments })
   }
 
-  const assignmentGroupedLevel: LevelAssignments = {}
-  assignmentsFiltered.forEach((assignment) => {
-    const { level } = assignment
-    if (!assignmentGroupedLevel[level]) assignmentGroupedLevel[level] = [assignment]
-    else assignmentGroupedLevel[level].push(assignment)
-  })
+  const defaultAssignedSections = Object.values(fullStructure.sections).reduce(
+    (assignedSections, { details: { code } }) => ({
+      ...assignedSections,
+      [code]: { newAssignee: undefined },
+    }),
+    {}
+  )
 
-  const currentReviewLevel = Math.max(Number(Object.keys(assignmentGroupedLevel)))
-
-  const isFullyAssigned = currentReviewLevel
-    ? calculateIsFullyAssigned(assignmentGroupedLevel[currentReviewLevel], sectionCodes)
-    : true
+  const AssignmentSection: React.FC<SectionDetails> = ({ id, title, code }) => {
+    const { reviewStructuresState } = useReviewStructureState()
+    return (
+      <Segment key={id}>
+        <Header className="section-title" as="h5" content={title} />
+        {Object.entries(assignmentGroupedLevel).map(([level, assignments]) => (
+          <div className="flex-column" key={`assignment-group-level-${level}`}>
+            {fullStructure.info.outcome === ApplicationOutcome.Pending && (
+              <AssignmentSectionRow
+                assignments={assignments}
+                sectionCode={code}
+                reviewLevel={Number(level)}
+                structure={fullStructure}
+                assignedSections={assignedSectionsByLevel[level] || defaultAssignedSections}
+                setAssignedSections={(assignedSections) =>
+                  setAssignedSectionsByLevel({
+                    ...assignedSectionsByLevel,
+                    [level]: assignedSections,
+                  })
+                }
+                setEnableSubmit={setEnableSubmit}
+                setAssignmentError={setAssignmentError}
+              />
+            )}
+            {(assignments as AssignmentDetails[]).map((assignment) =>
+              reviewStructuresState[assignment.id] && assignment.isCurrentUserReviewer ? (
+                <ReviewSectionRow
+                  key={`review-row-section-${code}-assignment-${assignment.id}`}
+                  sectionId={id}
+                  fullStructure={fullStructure}
+                  reviewAssignment={assignment}
+                  previousAssignment={assignmentInPreviousStage}
+                />
+              ) : null
+            )}
+          </div>
+        ))}
+      </Segment>
+    )
+  }
 
   return (
     <Container id="assignment-tab">
@@ -131,36 +140,25 @@ const AssignmentTab: React.FC<{
         />
       )}
       <div className="flex-row-space-between-center" id="review-filters-container">
-        <ReviewLevel
-          filters={filters}
-          setFilters={setFilters}
-          structure={fullStructure}
-          assignments={assignments}
-        />
+        <ReviewLevel filters={filters} setFilters={setFilters} structure={fullStructure} />
         <div className="centered-flex-box-row">
           <Label className="uppercase-label" content={strings.REVIEW_OVERVIEW_STAGE} />
           <Stage name={stageName} colour={stageColour || ''} />
         </div>
       </div>
-      {assignmentsFiltered.length > 0 && (
+      {loading ? (
+        <Loading />
+      ) : !assignmentsFiltered || assignmentsFiltered.length === 0 ? (
+        <NoMatch />
+      ) : (
         <ReviewStateProvider assignments={assignmentsFiltered}>
-          <AssignmentRows
-            fullStructure={fullStructure}
-            assignmentInPreviousStage={assignmentInPreviousStage}
-            assignmentGroupedLevel={assignmentGroupedLevel}
-            assignedSectionsByLevel={assignedSectionsByLevel}
-            setAssignedSectionsByLevel={setAssignedSectionsByLevel}
-            setEnableSubmit={setEnableSubmit}
-            setAssignmentError={setAssignmentError}
-          />
+          {Object.values(fullStructure.sections).map(({ details }) => (
+            <AssignmentSection {...details} />
+          ))}
         </ReviewStateProvider>
       )}
       {!isFullyAssigned && (
-        <AssignAll
-          assignments={assignmentsFiltered}
-          setReviewerForAll={assignAllSections}
-          currentUser={currentUser}
-        />
+        <AssignAll assignments={assignmentsFiltered} setReviewerForAll={assignAllSections} />
       )}
       {fullStructure.info.outcome === ApplicationOutcome.Pending && (
         <AssignmentSubmit
@@ -169,7 +167,6 @@ const AssignmentTab: React.FC<{
           assignmentsFiltered={assignmentsFiltered}
           enableSubmit={enableSubmit}
           setAssignmentError={setAssignmentError}
-          refetch={refetch}
         />
       )}
     </Container>
@@ -177,17 +174,3 @@ const AssignmentTab: React.FC<{
 }
 
 export default AssignmentTab
-
-const calculateIsFullyAssigned = (
-  currentLevelAssignments: AssignmentDetails[],
-  sectionCodes: string[]
-) => {
-  const assignedSections = new Set(
-    currentLevelAssignments
-      // .filter((assignment) => assignment.current.assignmentStatus === ReviewAssignmentStatus.Assigned)
-      .map((assignment) => assignment.assignedSections)
-      .flat()
-  )
-
-  return assignedSections.size >= sectionCodes.length
-}
